Fixes error when trying to sync large directories
The problem was that the remote host rejected incomming connections.
Right now the code is written in a way that all files that should be
synced are identified put into a list and that list will then be
evaluated. Unfortunately the plugin tried to open a new SFTP connection
for every file. And since every syncing process runs in its own Promise
the host got swamped by connection requests.

The solution to that at the moment is to throttle/pool the promises and
thus the connections to the remote host.